Creates a registration mechanism for FSM commands.
